+++ src/NewPicker/PickerInput/hooks/useRangePickerValue.ts
@@ -3,7 +3,7 @@ import useLayoutEffect from 'rc-util/lib/hooks/useLayoutEffect';
 import * as React from 'react';
 import type { GenerateConfig } from '../../../generate';
 import { isSame } from '../../../utils/dateUtil';
-import type { InternalMode, Locale, PanelMode } from '../../interface';
+import type { InternalMode, Locale, PanelMode } from '../../../interface';
 import type { RangePickerProps } from '../RangePicker';
 
 export function offsetPanelDate<DateType = any>(
