+++ src/NewPicker/PickerPanel/DatePanel/index.tsx
@@ -7,7 +7,7 @@ import {
   isSameMonth,
   WEEK_DAY_COUNT,
 } from '../../../utils/dateUtil';
-import type { PanelMode, SharedPanelProps } from '../../interface';
+import type { PanelMode, SharedPanelProps } from '../../../interface';
 import { PanelContext, useInfo } from '../context';
 import PanelBody from '../PanelBody';
 import PanelHeader from '../PanelHeader';
